+++ exchange/src/exchange/algos/sandbox_mode/sandbox_mode.cpp
@@ -1,4 +1,5 @@
 #include "sandbox_mode.hpp"
+
 #include "exchange/config.h"
 
 #include <glaze/glaze.hpp>
