+++ exchange/src/exchange/bots/bot_types/retail.hpp
@@ -1,20 +1,20 @@
 #pragma once
 
-#include "generic_bot.hpp"
+#include "exchange/traders/trader_types/bot_trader.hpp"
 #include "shared/messages_wrapper_to_exchange.hpp"
 
 namespace nutc {
 
 namespace bots {
 
-class RetailBot : public GenericBot {
+class RetailBot : public BotTrader {
     std::random_device rd{};
     std::mt19937 gen{};
     std::poisson_distribution<> poisson_dist{};
 
 public:
     RetailBot(std::string ticker, double interest_limit) :
-        GenericBot(std::move(ticker), interest_limit),
+        BotTrader(std::move(ticker), interest_limit),
         AGGRESSIVENESS(std::normal_distribution<>{50, 2000}(gen))
     {
         set_capital(interest_limit);
@@ -23,7 +23,10 @@ public:
     [[nodiscard]] bool is_active() const override;
 
     std::optional<messages::MarketOrder> take_action(double current, double theo);
-    RetailBot(RetailBot&& other) noexcept : GenericBot(std::move(other)), AGGRESSIVENESS(other.AGGRESSIVENESS) {}
+
+    RetailBot(RetailBot&& other) noexcept :
+        BotTrader(std::move(other)), AGGRESSIVENESS(other.AGGRESSIVENESS)
+    {}
 
 private:
     static double calculate_order_price(
