+++ exchange/src/exchange/process_spawning/spawning.cpp
@@ -1,7 +1,6 @@
 #include "spawning.hpp"
 
 #include "exchange/logging.hpp"
-#include "exchange/traders/trader_types.hpp"
 #include "shared/file_operations/file_operations.hpp"
 
 #include <cstdlib>
