+++ exchange/src/exchange/traders/trader_types/trader_type.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace nutc {
+namespace manager {
+enum TraderType { REMOTE, LOCAL, BOT };
+}
+} // namespace nutc
